refactor: split large components.ts into category-specific modules
- Create form-components.ts for form and input interfaces (93 lines)
- Create layout-components.ts for layout and navigation interfaces (108 lines)
- Create feedback-components.ts for feedback and notification interfaces (69 lines)
- Create data-components.ts for data display and presentation interfaces (116 lines)
- Create utility-components.ts for utility and generic interfaces (110 lines)
- Update components.ts to re-export from focused modules (77 lines)

This addresses the code review feedback about the 470-line components.ts
file being too large. The new structure organizes component prop interfaces
by category (form, layout, feedback, data, utility) while preserving
backward compatibility through re-exports.
